+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Converted `SlackCredentials` and `SlackWebhook` into a `Block` - [#10](https://github.com/PrefectHQ/prefect-slack/pull/10)
+- Converted `SlackCredentials` and `SlackWebhook` into a `Block` - [#29](https://github.com/PrefectHQ/prefect-slack/pull/29)
 
 
 ### Deprecated